refactor: change components names
